+++ cmd/konvoy-image/cmd/validate.go
@@ -1,9 +1,6 @@
 package cmd
 
 import (
-	"fmt"
-	"net"
-
 	"github.com/pkg/errors"
 	"github.com/spf13/cobra"
 
@@ -18,18 +15,10 @@ var validateCmd = &cobra.Command{
 	SilenceErrors: true,
 	Use:           "validate",
 	Short:         "validate existing infrastructure",
-	Args:          cobra.ExactArgs(1),
+	Args:          cobra.ExactArgs(0),
 	RunE: func(cmd *cobra.Command, args []string) error {
-		if !isValidCIDR(validateFlags.PodSubnet) {
-			return fmt.Errorf("pod-subnet %q was not a valid CIDR", validateFlags.PodSubnet)
-		}
-
-		if !isValidCIDR(validateFlags.ServiceSubnet) {
-			return fmt.Errorf("service-subnet %q was not a valid CIDR", validateFlags.ServiceSubnet)
-		}
-
-		if err := app.Validate(args[0], validateFlags); err != nil {
-			return errors.Wrap(err, "error running provision")
+		if err := app.Validate(validateFlags); err != nil {
+			return errors.Wrap(err, "error running validate")
 		}
 
 		return nil
@@ -40,15 +29,10 @@ func init() {
 	rootCmd.AddCommand(validateCmd)
 
 	flagSet := validateCmd.Flags()
+	flagSet.StringVar(&validateFlags.Inventory, "inventory-file", "inventory.yaml", "an ansible inventory defining your infrastructure")
 	flagSet.StringVar(&validateFlags.ServiceSubnet, "service-subnet", "10.96.0.0/12", "ip addresses used"+
 		" for the service subnet")
 	flagSet.StringVar(&validateFlags.PodSubnet, "pod-subnet", "192.168.0.0/16", "ip addresses used"+
 		" for the pod subnet")
 	flagSet.IntVar(&validateFlags.APIServerPort, "apiserver-port", 6443, "apiserver port")
-}
-
-func isValidCIDR(str string) bool {
-	_, _, err := net.ParseCIDR(str)
-
-	return err == nil
 }
