+++ pkg/app/provision.go
@@ -1,16 +1,10 @@
 package app
 
 import (
-	"fmt"
-
 	"github.com/pkg/errors"
 
 	"github.com/mesosphere/konvoy-image-builder/pkg/ansible"
 	"github.com/mesosphere/konvoy-image-builder/pkg/appansible"
-)
-
-const (
-	extraVarsTemplate = "%s=%s"
 )
 
 type ProvisionFlags struct {
@@ -22,26 +16,6 @@ type ProvisionFlags struct {
 	Provider  string
 	WorkDir   string
 	Inventory string
-}
-
-type ValidateFlags struct {
-	RootFlags
-
-	ServiceSubnet string
-	PodSubnet     string
-	APIServerPort int
-}
-
-func Validate(inventory string, flags ValidateFlags) error {
-	playbookOpts := validateFlagsToPlaybookOptions(flags)
-
-	playbook := appansible.NewPlaybook("validate", inventory, playbookOpts)
-
-	if err := playbook.Run(NewRunOptions(flags.RootFlags)); err != nil {
-		return errors.Wrap(err, "error running playbook")
-	}
-
-	return nil
 }
 
 func Provision(inventory string, flags ProvisionFlags) error {
@@ -63,21 +37,4 @@ func NewRunOptions(flags RootFlags) appansible.RunOptions {
 		RunsDirectory: AnsibleRunsDirectory,
 		Verbosity:     flags.Verbosity,
 	}
-}
-
-func validateFlagsToPlaybookOptions(flags ValidateFlags) *ansible.PlaybookOptions {
-	playbookOptions := &ansible.PlaybookOptions{}
-	values := []string{
-		fmt.Sprintf(extraVarsTemplate, "service_subnet", flags.ServiceSubnet),
-		fmt.Sprintf(extraVarsTemplate, "pod_subnet", flags.PodSubnet),
-		fmt.Sprintf("%s=%d", "apiserver_port", flags.APIServerPort),
-	}
-
-	if playbookOptions.ExtraVars == nil {
-		playbookOptions.ExtraVars = make([]string, 0)
-	}
-
-	playbookOptions.ExtraVars = append(playbookOptions.ExtraVars, values...)
-
-	return playbookOptions
 }
